arrays: separate runtime concerns
